Working implementations for PCM_16, PCM_32, Float_32 and Float_64 (With some bugs)
Commit contains the code for the minimal working project. Allows for
reading and writing of the types of wav file mentioned in the title.
Each type can be easily converted to another and the changes are
reflected then when writing the file to disk.

The minor bugs: Namely a segfault when reading 32-bit floats in bulk.
Individual files can be read succesfully but for some reason when the
code is used in PyWavers to read in a list of wav files a segfault
occurs.