Removal of ndarrays and start of a rewrite of SampleType
Benchmarking indicated that a significant amount of time. Changed the
interface to return a vector of channel interleaved samples. Added a
script to get some of GenSpeech for benching purposes

Updated all the tests to match the interface change. This involved
re-writing re-writing the test cases, which I was planning anyways, but
I was a bit lazy with creation and removal of test files and now it
needs to be run as ``cargo test -- --test-threads 1`` to avoid some
parallel files accessing :( Will fix v soon

Started a re-write of the SampleType enum. It felt very messy and was a
bit of an eye-sore so I am cleaning it up.