V0.2.0 - Ndarrays are now a feature + Sample refactor
Big changes to the interface. Any where that used to return an
Array2 now returns a vec of interleaved samples instead.

The performance difference was not as big as previously thought between
vec and ndarray. Instead it was a poor implementation that resulted in
another iteration of the entire signal.

Addressed this performance issue simply by adding in a check to see if
the type desired by the user is actually different then how the wave
file is actually represented. If the are the same then the function
returns immediately.

Introduces the IntoArray trait which is designed to be used to convert
a  Vec<T>, namely T being Sample, i16, i32, f32 and f64, to an
Array2<T>.

The into_array function consumes the vector and requires the number of
channels in the signal to be passed so it can correctly structure the
signal.

Enable the IntoArray trait by building with the ``ndarray`` feature
flag.

Finished the refactor of the ``Sample`` enum. Mostly it is just a name
change to be clearer and the same with the interface for converting
between sample types.

Alongside the ``Sample`` refactor, I introduced the AudioConversion and IterAudioConversion
traits for converting between audio sample representations for
individual samples and vectors of samples.

Removed/changed around some dependencies. Removed rayon and itertools as
they are not necessary at the moment.