Complete Re-write
Ignoring best practices, this commit is large and won't detail all
that's changed since the previous coommit.

Contained in this commit is a complete rewrite of WaveRs from the
ground up. This time, taking the lessons from the first attempt. The
goal with the rewrite was to A) improve ease of use. The old API felt
clunky and (B) improve performance. Performance was always a key goal of
the project but unfortunately didn't materialise in the first iteration.
However, according to the benchmarks in both Rust and Python, an
acceptable level of performance has been achieved.

The library is now better tested and benchmarked too.

The README has been given a face lift and so too have the docs.