Wavers 1.1.0
Updated Wavers to v1.1.0. This is a significant update to the library
which involved nearly a full rewrite. There were several motivating
factors which prompted this.

1) Wanting to provide a more easy-to-use interface for the library. Now
   a user does not need to read an entire wav file. They can, or they
can create a ``Wav`` struct which can later be read from. The struct
contains information on the wav file such as sample rate, channels,
encoding, etc.

2) Remove any use of unstable, nightly features. The main one being
   const type ids. This rewrite removes the need for them, while also
maintaining the same functionality.

3) Wanting to tidy up the functionality around reading. Previously it
   felt overly verbose. It might not be any shorter now, but it is
better laid out and more reliant on concrete types.

4) Converting between encodings was not as nice as I would like and
   definitely prone to bugs. You could specify a type ``T`` of a Wav
file and it would try and read the wav file as that type. This is
incorrect. Now, the native encoding is first checked, the samples are
loaded as this native encoding and then  converted to the desired type.

5) Wanted to clean up/improve the conversion functionality. Now there is
   a ConvertSlice trait which handles most of the conversion.