Bug Fix: Wav::n_samples() returning incorrect values after conversion
Fixed a bug where the n_samples function would return an incorrect
number. This was caused by using the size of the conversion type to
calculate the number of samples by dividing native size (size before
conversion) by the size of the conversion type.

Now the n_samples function does not calculate the number of samples
using the conversion type at all and instead uses the native information
from the wav header.

Added some tests to verify the fix