PCM 24-bit support
Added an ``i24`` type which is used to read wav files encoded in the
PCM_24 format. Under the hood it is a byte array. For most operations
it is converted to an ``i32``, operated on, and then converted back to
an ``i24``. Added two unit tests for this. Data was taken as always from
SoundFile and Numpy.