Possibly working implementation of Extensible Wav Format
Commit contains the (hopefully) necessary implemenation details to allow
the reading and writing of the extesible wav format. It passes all of
the current tests with the modifications. However, the current tests do
not test the extensible format.

Some changes to behaviour:
	1) The in-memory respresentation of the format chunk is now 40
	   bytes, compared to 16 bytes previously. The additional 24
bytes are used to represent the extensible format fields.
	2) Non-PCM data, i.e. floating point is now assumed to be the
	   extensible format. That is, wavers now writes float wav files
in the extensible format. NOTE: We can still read non-extensible format
float files.
	3) We only write what is necessary, so if the file is in PCM
	   format, it will be written as 16 bytes instead of 40.

Next commit will be new tests. Lots of cleanup to do too I feel