Extensible Format, FACT chunk and LIST chunk + MORE
Moving closer to finalising the work on the extensible format. Should be
fully readable and writable now. Added explicit support for retrieving
the FACT and LIST chunks.

They are accessible via the ``Wav::get_fact_chunk`` and
``Wav::get_list_chunk`` functions.

Introduced the ``Chunk`` trait that is intended to be used with all
chunks except for the DATA and RIFF chunks. These are excluded (at least
for now) since they operate slightly differently when compared to the
other chunks. Again at least for now. The trait provide a means of
decoding, encoding and displaying chunks in a uniform way. Each
``Chunk`` implementation must include the ``id``, ``size``,
``as_bytes``, and ``from_reader`` functions.  The first three are self
explanatory and the ``from_reader`` function takes in a reader and is
used when decoding the chunks using the ``reader`` field of the ``Wav``
struct.

The ``Chunk`` trait comes alongside the ``read_chunk`` function which is
allows for the generic creation of any ``Chunk`` implementaton. See
``Wav::get_chunk`` and ``Wav::get_list/fact_chunk`` for a reference.

Refactored the structure of the project a bit. Moved all the chunks into
``chunks`` module.

Refactored the file size calculation to loop over the ``header_info``
map of the ``WavHeader`` struct and sum the values. This comes alongside
an internal change to store the offset as the start of a chunk, not the
start of the data of the chunk (usually just the chunk id and chunk size
are skipped). This was done to hopefully prevent confusion on my end.
Doens't affect anything on the top-level api and doesn't introduce any
functionality changes.

Colorised printing! Added the optional dependency on ``colored`` to
enable slightly prettier printing of wav information to the console. If
enabled, when displaying the chunks and header information it will be
printed in a prettier way. The ``Chunk`` trait requires that the
implementing struct also implements the ``Display`` trait. Stay tuned
for the ``wav_header`` or maybe ``wav_inspect`` tool for printing all
the chunk information of a given wav file. Just need to include this
crate as a dependency on it in such a way that it can make use of these
changes without having to push this to crates.

While some have tested the new functionality relating to the extensible
format, it is still fairly untested. All the current tests are still
passing, even after they were updated to write the likes of the IEEE
float format in the extensible format so that's a good sign.

More polishing to come before merging with main and publishing