Frame and Channel Iterators
Adds two iterators that can be created for Wav struct; the FrameIterator
which iterates over Frames of the wav file. A frame is defined as a set
of samples across all channels for a single sample point. The
ChannelIterator then iterates of each channel.

Wavers normally reads all data at once and returns return the channel
data as an interleaved slice. Now the FrameIterator allows the reading
of individual frames at a time. For example, in a two channel wav file,
``samples[0]``refers to the first sample in channel 1, and
``samples[1]`` refers to the first samples in channel 2, and so on. With
the frame iterator for the same file, each iteration of the iterator
will yield a ``Samples`` struct with a length of two. It is the
equivalent of calling ``Wav.read_samples(n_channels)`` repeatedly.

The ``ChannelIterator`` then allows to access each channel at a time. It
returns a ``Samples`` struct containing an entire channel and the next
iteration will yield the next channel.

See the README for examples of both.