[#19] Streamlined Benchmarks
Have yet to run the full suite. It takes time :(

Complete re-wrote the benchmarks so they are simpler and more direct in
their measurements. Currently, there are five benchmarks which should
provide a broad overview of the performance of the crate.

1. Sample Conversion - How long it takes to convert between the
   various sample types?
2. Wavers Reading - How long do it take for wavers to read a wav file?
3. Hound Reading - Hound is the most similar crate. How long does it
   take to read a wav file?
4. Wavers Writing - How long does it take wavers to write a file?
5. Hound Writing - How long does it take hound to write a file?

These reading and writing benchmarks cover reading and writing the same
types (no conversion takes place) for all types. Exception being with
Hound and i24 and f64 since it does not support them.

Implemented using a few macros.