Fixed Incorrect Size Issues and Playback Issues
Fixed two issues from GitHub. #23 #24

There was a mismatch between the actual file size and the reported file
size caused by an extra addition when creating headers.

Then the playback issue was fixed because it seems a lot of players
don't support the extensible format. Changed it so the float format
simple uses the float format instead of the extensible format.

All tests pass. Introduced a new test for the size issue and it passes.
Testing the reproducible code from both issues and have fixed both.

Changed the crate::write function to use a buf writer instead of file writing since there are multiple repeated writes.